+++ src/sorter.rs
@@ -12,14 +12,9 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-use rayon::prelude::*;
-use std::{
-    cmp::Ordering,
-    collections::VecDeque,
-    fs::{File, OpenOptions},
-    io::{BufReader, BufWriter, Error, Read, Seek, SeekFrom, Write},
-    path::{Path, PathBuf},
-};
+use std::{cmp::Ordering, io::Error, path::PathBuf};
+
+use crate::{iter::SortedIterator, push::PushExternalSorter, Sortable};
 
 /// Exposes external sorting (i.e. on disk sorting) capability on arbitrarily
 /// sized iterator, even if the generated content of the iterator doesn't fit in
@@ -75,272 +70,82 @@ impl ExternalSorter {
 
     /// Sorts a given iterator, returning a new iterator with items
     pub fn sort<T, I>(
-        &self,
+        self,
         iterator: I,
     ) -> Result<SortedIterator<T, impl Fn(&T, &T) -> Ordering + Send + Sync>, Error>
     where
         T: Sortable + Ord,
-        I: Iterator<Item = T>,
+        I: IntoIterator<Item = T>,
     {
         self.sort_by(iterator, |a, b| a.cmp(b))
     }
 
-    /// Sorts a given iterator with a comparator function, returning a new iterator with items
-    pub fn sort_by<T, I, F>(&self, iterator: I, cmp: F) -> Result<SortedIterator<T, F>, Error>
-    where
-        T: Sortable,
-        I: Iterator<Item = T>,
-        F: Fn(&T, &T) -> Ordering + Send + Sync,
-    {
-        let mut tempdir: Option<tempfile::TempDir> = None;
-        let mut sort_dir: Option<PathBuf> = None;
-
-        let mut count = 0;
-        let mut segments_file: Vec<File> = Vec::new();
-        let mut buffer: Vec<T> = Vec::with_capacity(self.segment_size);
-        for next_item in iterator {
-            count += 1;
-            buffer.push(next_item);
-            if buffer.len() > self.segment_size {
-                let sort_dir = self.lazy_create_dir(&mut tempdir, &mut sort_dir)?;
-                self.sort_and_write_segment(sort_dir, &mut segments_file, &mut buffer, &cmp)?;
-            }
-        }
-
-        // Write any items left in buffer, but only if we had at least 1 segment
-        // written. Otherwise we use the buffer itself to iterate from memory
-        let pass_through_queue = if !buffer.is_empty() && !segments_file.is_empty() {
-            let sort_dir = self.lazy_create_dir(&mut tempdir, &mut sort_dir)?;
-            self.sort_and_write_segment(sort_dir, &mut segments_file, &mut buffer, &cmp)?;
-            None
-        } else {
-            buffer.sort_unstable_by(&cmp);
-            Some(VecDeque::from(buffer))
-        };
-
-        SortedIterator::new(tempdir, pass_through_queue, segments_file, count, cmp)
-    }
-
     /// Sorts a given iterator with a key extraction function, returning a new iterator with items
     pub fn sort_by_key<T, I, F, K>(
-        &self,
+        self,
         iterator: I,
         f: F,
     ) -> Result<SortedIterator<T, impl Fn(&T, &T) -> Ordering + Send + Sync>, Error>
     where
         T: Sortable,
-        I: Iterator<Item = T>,
-        F: Fn(&T) -> K + Send + Sync,
+        I: IntoIterator<Item = T>,
+        F: Fn(&T) -> K + Send + Sync + Clone,
         K: Ord,
     {
         self.sort_by(iterator, move |a, b| f(a).cmp(&f(b)))
     }
 
-    /// We only want to create directory if it's needed (i.e. if the dataset
-    /// doesn't fit in memory) to prevent filesystem latency
-    fn lazy_create_dir<'a>(
-        &self,
-        tempdir: &mut Option<tempfile::TempDir>,
-        sort_dir: &'a mut Option<PathBuf>,
-    ) -> Result<&'a Path, Error> {
-        if let Some(sort_dir) = sort_dir {
-            return Ok(sort_dir);
-        }
-
-        *sort_dir = if let Some(ref sort_dir) = self.sort_dir {
-            Some(sort_dir.to_path_buf())
-        } else {
-            *tempdir = Some(tempfile::TempDir::new()?);
-            Some(tempdir.as_ref().unwrap().path().to_path_buf())
-        };
-
-        Ok(sort_dir.as_ref().unwrap())
-    }
-
-    fn sort_and_write_segment<T, F>(
-        &self,
-        sort_dir: &Path,
-        segments: &mut Vec<File>,
-        buffer: &mut Vec<T>,
-        cmp: F,
-    ) -> Result<(), Error>
+    /// Sorts a given iterator with a comparator function, returning a new iterator with items
+    pub fn sort_by<T, I, F>(self, iterator: I, cmp: F) -> Result<SortedIterator<T, F>, Error>
     where
         T: Sortable,
-        F: Fn(&T, &T) -> Ordering + Send + Sync,
+        I: IntoIterator<Item = T>,
+        F: Fn(&T, &T) -> Ordering + Send + Sync + Clone,
     {
-        if self.parallel {
-            buffer.par_sort_unstable_by(|a, b| cmp(a, b));
-        } else {
-            buffer.sort_unstable_by(|a, b| cmp(a, b));
-        }
+        let mut sorter =
+            PushExternalSorter::new(self.segment_size, self.sort_dir, self.parallel, cmp);
+        sorter.push_iter(iterator)?;
+        sorter.done()
+    }
 
-        let segment_path = sort_dir.join(format!("{}", segments.len()));
-        let segment_file = OpenOptions::new()
-            .create(true)
-            .truncate(true)
-            .read(true)
-            .write(true)
-            .open(segment_path)?;
-        let mut buf_writer = BufWriter::new(segment_file);
+    /// Creates a pushed external sorter, which will consume items in a push
+    /// pattern and compare them using the default comparator.
+    pub fn pushed<T>(
+        self,
+    ) -> PushExternalSorter<T, impl Fn(&T, &T) -> Ordering + Send + Sync + Clone>
+    where
+        T: Sortable + Ord,
+    {
+        self.pushed_by::<T, _>(|a, b| a.cmp(b))
+    }
 
-        for item in buffer.drain(0..) {
-            item.encode(&mut buf_writer);
-        }
+    /// Creates a pushed external sorter, which will consume items in a push
+    /// pattern and compare them using the given comparator function.
+    pub fn pushed_by<T, F>(self, cmp: F) -> PushExternalSorter<T, F>
+    where
+        T: Sortable,
+        F: Fn(&T, &T) -> Ordering + Send + Sync + Clone,
+    {
+        PushExternalSorter::new(self.segment_size, self.sort_dir, self.parallel, cmp)
+    }
 
-        let file = buf_writer.into_inner()?;
-        segments.push(file);
-
-        Ok(())
+    /// Creates a pushed external sorter, which will consume items in a push
+    /// pattern and compare them using the given key extraction function.
+    pub fn pushed_by_key<T, F, K>(
+        self,
+        f: F,
+    ) -> PushExternalSorter<T, impl Fn(&T, &T) -> Ordering + Send + Sync + Clone>
+    where
+        T: Sortable,
+        F: Fn(&T) -> K + Send + Sync + Clone,
+        K: Ord,
+    {
+        self.pushed_by(move |a, b| f(a).cmp(&f(b)))
     }
 }
 
 impl Default for ExternalSorter {
     fn default() -> Self {
         ExternalSorter::new()
-    }
-}
-
-pub trait Sortable: Sized + Send {
-    fn encode<W: Write>(&self, writer: &mut W);
-    fn decode<R: Read>(reader: &mut R) -> Option<Self>;
-}
-
-pub struct SortedIterator<T: Sortable, F> {
-    _tempdir: Option<tempfile::TempDir>,
-    pass_through_queue: Option<VecDeque<T>>,
-    segments_file: Vec<BufReader<File>>,
-    next_values: Vec<Option<T>>,
-    count: u64,
-    cmp: F,
-}
-
-impl<T: Sortable, F: Fn(&T, &T) -> Ordering + Send + Sync> SortedIterator<T, F> {
-    fn new(
-        tempdir: Option<tempfile::TempDir>,
-        pass_through_queue: Option<VecDeque<T>>,
-        mut segments_file: Vec<File>,
-        count: u64,
-        cmp: F,
-    ) -> Result<SortedIterator<T, F>, Error> {
-        for segment in &mut segments_file {
-            segment.seek(SeekFrom::Start(0))?;
-        }
-
-        let next_values = segments_file
-            .iter_mut()
-            .map(|file| T::decode(file))
-            .collect();
-
-        let segments_file_buffered = segments_file.into_iter().map(BufReader::new).collect();
-
-        Ok(SortedIterator {
-            _tempdir: tempdir,
-            pass_through_queue,
-            segments_file: segments_file_buffered,
-            next_values,
-            count,
-            cmp,
-        })
-    }
-
-    pub fn sorted_count(&self) -> u64 {
-        self.count
-    }
-}
-
-impl<T: Sortable, F: Fn(&T, &T) -> Ordering> Iterator for SortedIterator<T, F> {
-    type Item = T;
-
-    fn next(&mut self) -> Option<T> {
-        // if we have a pass through, we dequeue from it directly
-        if let Some(ptb) = self.pass_through_queue.as_mut() {
-            return ptb.pop_front();
-        }
-
-        // otherwise, we iter from segments on disk
-        let mut smallest_idx: Option<usize> = None;
-        {
-            let mut smallest: Option<&T> = None;
-            for idx in 0..self.segments_file.len() {
-                let next_value = self.next_values[idx].as_ref();
-                if next_value.is_none() {
-                    continue;
-                }
-
-                if smallest.is_none()
-                    || (self.cmp)(next_value.unwrap(), smallest.unwrap()) == Ordering::Less
-                {
-                    smallest = Some(next_value.unwrap());
-                    smallest_idx = Some(idx);
-                }
-            }
-        }
-
-        smallest_idx.map(|idx| {
-            let file = &mut self.segments_file[idx];
-            let value = self.next_values[idx].take().unwrap();
-            self.next_values[idx] = T::decode(file);
-            value
-        })
-    }
-}
-
-#[cfg(test)]
-pub mod test {
-    use super::*;
-
-    use byteorder::{ReadBytesExt, WriteBytesExt};
-
-    #[test]
-    fn test_smaller_than_segment() {
-        let sorter = ExternalSorter::new();
-        let data: Vec<u32> = (0..100u32).collect();
-        let data_rev: Vec<u32> = data.iter().rev().cloned().collect();
-
-        let sorted_iter = sorter.sort(data_rev.into_iter()).unwrap();
-
-        // should not have used any segments (all in memory)
-        assert_eq!(sorted_iter.segments_file.len(), 0);
-        let sorted_data: Vec<u32> = sorted_iter.collect();
-
-        assert_eq!(data, sorted_data);
-    }
-
-    #[test]
-    fn test_multiple_segments() {
-        let sorter = ExternalSorter::new().with_segment_size(100);
-        let data: Vec<u32> = (0..1000u32).collect();
-
-        let data_rev: Vec<u32> = data.iter().rev().cloned().collect();
-        let sorted_iter = sorter.sort(data_rev.into_iter()).unwrap();
-        assert_eq!(sorted_iter.segments_file.len(), 10);
-
-        let sorted_data: Vec<u32> = sorted_iter.collect();
-        assert_eq!(data, sorted_data);
-    }
-
-    #[test]
-    fn test_parallel() {
-        let sorter = ExternalSorter::new()
-            .with_segment_size(100)
-            .with_parallel_sort();
-        let data: Vec<u32> = (0..1000u32).collect();
-
-        let data_rev: Vec<u32> = data.iter().rev().cloned().collect();
-        let sorted_iter = sorter.sort(data_rev.into_iter()).unwrap();
-        assert_eq!(sorted_iter.segments_file.len(), 10);
-
-        let sorted_data: Vec<u32> = sorted_iter.collect();
-        assert_eq!(data, sorted_data);
-    }
-
-    impl Sortable for u32 {
-        fn encode<W: Write>(&self, writer: &mut W) {
-            writer.write_u32::<byteorder::LittleEndian>(*self).unwrap();
-        }
-
-        fn decode<R: Read>(reader: &mut R) -> Option<u32> {
-            reader.read_u32::<byteorder::LittleEndian>().ok()
-        }
     }
 }